Fix element-template-generator to support long version numbers
- Change VersionedElementTemplate.version from Integer to Long
- Change ElementTemplate.version from int to long
- Update getBasicElementTemplate to return Optional and skip unparseable templates
- Templates with version > Integer.MAX_VALUE (like Web Modeler's epoch timestamps) now work

Agent-Logs-Url: https://github.com/camunda/connectors/sessions/c1d29290-f1dc-4856-84b6-21609227951c

+++ element-template-generator/core/src/main/java/io/camunda/connector/generator/dsl/ElementTemplate.java
@@ -46,7 +46,7 @@ import java.util.Set;
 public record ElementTemplate(
     String id,
     String name,
-    int version,
+    long version,
     String documentationRef,
     Engines engines,
     String description,
